requirements.txt
Flask>=3.0.0
pdf2docx>=0.5.6
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.0.0
camelot-py>=0.11.0
ghostscript>=0.7
Werkzeug>=3.0.0